Remove the ability to install packages in the buildroot.
With eos7, the tree is not DPKG-based, and there is no longer a package
manager available at all. Remove all code and options related to that.

Later, we will define a separate OSTree output in eos-build-meta for
the image builder, so that we don't need to add ISO filesystem tools
into the user tree. That's tracked here:
<https://github.com/endlessm/eos-build-meta/issues/13>.