Run systemd-tmpfiles in the ostree checkout
This is to fix an issue during the 50-flatpak image hook,
which would fail at the appstream update with this:

    gi.repository.GLib.GError: g-io-error-quark: While pulling appstream2/x86_64 from remote flathub: open(O_TMPFILE): No such file or directory (1)

The root cause being a lack of `/var/tmp` in the ostree checkout.

It's surprising that things worked as well as they did until now, and
perhaps is due to the ostree containing more things in `/var/` than it
should.